Move scripts inside package for easier access
Also updates the installer to make resources and console
execution easier
