+++ setup.cfg
@@ -4,7 +4,7 @@ version = 1.0.0
 description = Alexa Teacher Models
 author = amazon-alexa 
 author_email = ssoltan@amazon.com
-url = https://www.github.com/https://github.com/alexa/alexa-teacher-models
+url = https://github.com/alexa/alexa-teacher-models
 license = Apache 2.0
 keywords=
     deep-learning
@@ -34,13 +34,16 @@ classifiers =
 
 [options]
 packages = find:
-
+include_package_data = True
 install_requires =
     numpy
     transformers
     sentencepiece>=0.1.95
-    torch>=1.7
+    torch>=1.7,<2.0
     protobuf<3.21.0
+
+[options.package_data]
+alexa_teacher_models = scripts/deepspeed/*.json
 
 [options.extras_require]
 repl:
